Refactor timezones, add import tests
